Move `get_auction` to utils
